+++ gh-pages/changelog.md
@@ -15,8 +15,13 @@ Changelog format adheres to [Keep a Changelog](http://keepachangelog.com/)</sub>
 
 ## [Unreleased]
 
+## [v0.10.1] - 2016-06-09
+
 - Replaced jshint with standardjs
 - fixed math formulas rendering in blog
+- few tests added
+- few doc improvements
+- using new static-props features
 
 ## [v0.10.0] - 2016-04-16